refactor: move player component into playbar, delete PlayerContextProvider
