+++ frontend/src/AppContextProvider.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from 'react';
+import React, { useState } from "react";
 import axios from "axios";
 
 const AppContext = React.createContext();
@@ -8,12 +8,15 @@ function AppContextProvider({ children }) {
     const [roomName, setRoomName] = useState(localStorage.getItem("roomName"));
     const [roomDesc, setRoomDesc] = useState();
     const [password, setPassword] = useState(localStorage.getItem("password"));
+
     const [userCount, setUserCount] = useState(0);
+
     const [playlist, setPlaylist] = useState([]);
+    const [playing, setPlaying] = useState(true);
+    const [currentSong, setCurrentSong] = useState();
+
     const [version, setVersion] = useState(false);
     const [socket, setSocket] = useState();
-    const [currentSong, setCurrentSong] = useState();
-    const versionRef = useRef(version);
     const [key, setKey] = useState(0);
 
     async function createRoom(room) {
@@ -22,7 +25,7 @@ function AppContextProvider({ children }) {
             room
         );
 
-        console.log(`Creating room...`)
+        console.log(`Creating room...`);
         localStorage.setItem("roomID", response.data._id);
         localStorage.setItem("roomName", response.data.name);
         localStorage.setItem("password", response.data.password);
@@ -31,29 +34,34 @@ function AppContextProvider({ children }) {
         setRoomDesc(response.data.description);
         setPassword(localStorage.getItem("password"));
     }
-    
+
     async function joinRoom(room) {
         const response = await axios.post(
             "http://localhost:3000/api/room/join/",
             room
         );
-        console.log(`Joining room...`)
+        console.log(`Joining room...`);
         if (response.data.name) {
             localStorage.setItem("roomID", response.data._id);
             localStorage.setItem("roomName", response.data.name);
             localStorage.setItem("password", response.data.password);
-            console.log(`[localStorage updated] room=${localStorage.getItem("roomName")}, pw=${localStorage.getItem("password")}, ID=${localStorage.getItem("roomID")}\n `);
+            console.log(
+                `[localStorage updated] room=${localStorage.getItem(
+                    "roomName"
+                )}, pw=${localStorage.getItem(
+                    "password"
+                )}, ID=${localStorage.getItem("roomID")}\n `
+            );
             setRoomID(localStorage.getItem("roomID"));
             setRoomName(localStorage.getItem("roomName"));
             setRoomDesc(response.data.description);
-            setPassword(localStorage.getItem("password"));   
+            setPassword(localStorage.getItem("password"));
 
             return "forward";
         } else {
             return response.data;
         }
     }
-
 
     // The context value that will be supplied to any descendants of this component.
     const context = {
@@ -75,9 +83,10 @@ function AppContextProvider({ children }) {
         setSocket,
         currentSong,
         setCurrentSong,
-        versionRef,
         key,
-        setKey
+        setKey,
+        playing,
+        setPlaying,
     };
 
     // Wraps the given child components in a Provider for the above context.
