+++ frontend/src/components/MusicPlayer/SongControls.js
@@ -1,11 +1,11 @@
 import React, { useState, useContext } from "react";
 import { Button, ButtonGroup, Container } from "react-bootstrap";
 import { FaStepBackward, FaPlay, FaPause, FaStepForward } from "react-icons/fa";
-import { PlayerContext } from "../../PlayerContextProvider";
+import { AppContext } from "../../AppContextProvider";
 import styles from "./SongControls.module.css";
 export default function SongControls() {
     // need context
-    const { playing, setPlaying } = useContext(PlayerContext);
+    const { playing, setPlaying } = useContext(AppContext);
     const [playButtonText, setPlayButtonText] = useState(<FaPause />);
 
     const nextSong = () => { };
